fixed file discrepancies
