+++ ieee123/ica_ieee123.glm
@@ -79,7 +79,7 @@ clock {
 #else
 #error TIMEZONE is not specified in ${CONFIGFILE}
 #endif // TIMEZONE
-module ica_internal_init;
+module ica_internal;
 
 ///////////////////////////////////////////////////////
 //
